feat(#38): replaced older launcher module
